+++ sparkle/types/objective.py
@@ -28,6 +28,7 @@ class SparkleObjective:
     minimise: bool
     post_process: typing.Callable
     use_time: UseTime
+    metric: bool
 
     def __init__(self: SparkleObjective,
                  name: str,
@@ -36,7 +37,8 @@ class SparkleObjective:
                  solver_aggregator: typing.Callable = None,
                  minimise: bool = True,
                  post_process: typing.Callable = None,
-                 use_time: UseTime = UseTime.NO) -> None:
+                 use_time: UseTime = UseTime.NO,
+                 metric: bool = False) -> None:
         """Create sparkle objective from string."""
         self.name = name
         self.run_aggregator: typing.Callable = run_aggregator
@@ -47,10 +49,16 @@ class SparkleObjective:
         self.minimise: bool = minimise
         self.post_process: typing.Callable = post_process
         self.use_time: UseTime = use_time
+        self.metric = metric
 
     def __str__(self: SparkleObjective) -> str:
         """Return a stringified version."""
-        return f"{self.name}"
+        return self.name
+
+    @property
+    def stem(self: SparkleObjective) -> str:
+        """Return the stem of the objective name."""
+        return self.name.split(":")[0]
 
     @property
     def time(self: SparkleObjective) -> bool:
@@ -66,7 +74,9 @@ class PAR(SparkleObjective):
                        SolverStatus.TIMEOUT,
                        SolverStatus.WRONG}
 
-    def __init__(self: PAR, k: int = 10) -> None:
+    def __init__(self: PAR, k: int = 10,
+                 minimise: bool = True,
+                 metric: bool = False) -> None:
         """Initialize PAR."""
         self.k = k
         if k <= 0:
@@ -78,4 +88,8 @@ class PAR(SparkleObjective):
                 return cutoff * self.k
             return value
 
-        super().__init__(f"PAR{k}", use_time=UseTime.CPU_TIME, post_process=penalise)
+        super().__init__(f"PAR{k}",
+                         minimise=minimise,
+                         use_time=UseTime.CPU_TIME,
+                         post_process=penalise,
+                         metric=metric)